Add and find comments for a Posting
Essentially a comment is a Posting, no need to create a new entity. Just add an
extra field call originalPostingId to posting, if this field is null, it means that
posting is top level and is not a comment, other wise posting is associated with
another post (i.e. posting with original posting Id not null is a comment).
Updated and extended tests.