Handle comments for postings
A comment is essentially a Posting entity with an extra field, here
defined as OriginalPostingId, to hold the Id of the Posting it is
a comment for.

Add repository methods to find comments for a Posting Id and
modified the getAll controller to only return top level
postings for cases where findAll() was previously invoked.

Modifed and extended tests.